Consolidate privacy utilities
